remove unix_getChildren, adjust COPYRIGHT info
