+++ src/java.base/unix/native/libjava/ProcessHandleImpl_unix.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2015, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -61,8 +61,6 @@ extern pid_t os_getParentPidAndTimings(JNIEnv *env, pid_t pid,
 
 extern void os_getCmdlineAndUserInfo(JNIEnv *env, jobject jinfo, pid_t pid);
 
-extern jint unix_getChildren(JNIEnv *env, jlong jpid, jlongArray array,
-                             jlongArray jparentArray, jlongArray jstimesArray);
 extern jint os_getChildren(JNIEnv *env, jlong jpid, jlongArray array,
                            jlongArray jparentArray, jlongArray jstimesArray);
